+++ codedigger/codeforces/tests/test_api.py
@@ -49,8 +49,4 @@ class TestAPI(TestSetUp):
         self.assertEqual(len(response), 10)
         self.assertRaises(ValidationException, user_status, wrong_handle)
 
-    def test_search_user(self):
-        url = reverse('search-user')
-        url += '?q=tou'
-        response = self.client.get(url, format = 'json')
-        self.assertEqual(response.status_code, 200)
+
