+++ codedigger/codeforces/tests/test_setup.py
@@ -2,6 +2,10 @@ from rest_framework.test import APITestCase
 
 
 class TestSetUp(APITestCase):
+    fixtures = [
+        "cf_users.json"
+    ]
+
     def setUp(self):
         return super().setUp()
